refactor: JpaRepository와 CustomRepository 상속받는 Repository 도입
