+++ src/main/java/roomescape/domain/ReservationRepository.java
@@ -1,25 +1,12 @@
 package roomescape.domain;
 
 import java.time.LocalDate;
-import java.util.List;
-import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReservationRepository {
-    List<Reservation> findAll(Long memberId, Long themeId, LocalDate dateFrom, LocalDate dateTo);
-
-    Optional<Reservation> findById(long id);
-
-    List<Long> findTimeIdByDateAndThemeId(LocalDate date, long themeId);
-
-    List<Theme> findThemeWithMostPopularReservation(String startDate, String endDate);
-
+public interface ReservationRepository extends JpaRepository<Reservation, Long>, ReservationCustomRepository {
     boolean existsByDateAndTimeIdAndThemeId(LocalDate date, long timeId, long themeId);
 
     boolean existsByTimeId(long timeId);
 
     boolean existsByThemeId(long themeId);
-
-    Reservation save(Reservation reservation);
-
-    void delete(Reservation reservation);
 }
