+++ src/main/java/roomescape/service/ReservationService.java
@@ -39,7 +39,7 @@ public class ReservationService {
 
     public List<ReservationResponse> findAllReservation(
             Long memberId, Long themeId, LocalDate dateFrom, LocalDate dateTo) {
-        List<Reservation> reservations = reservationRepository.findAll(memberId, themeId, dateFrom, dateTo);
+        List<Reservation> reservations = reservationRepository.searchAll(memberId, themeId, dateFrom, dateTo);
         return reservations.stream()
                 .map(ReservationResponse::new)
                 .toList();
